Create Googel Apps Script
I am trying to ease record problem for a startup that deals with book reviews. So, I created a database where the reviewers' submissions will be recorded so that payment status can be tracked easily. However, one person may submit multiple reviews, which could be classified as verified and unverified. Since the two types of review attract different amounts, I created this Google Apps Script to help me extract a list of unpaid reviewers classified according to the review type they do. This will help the owner to view the list of unpaid reviewers and make payment to them accordingly.